refactor date_time
